手动输表号、手动采集检表流程完成(但是容错能力较差)
git-svn-id: https://192.168.1.124:8443/svn/delusoft/trunk/qtverify@289 62fcab80-fd93-7748-8c49-e77283212eb3

+++ source/include/basedef.h
@@ -134,7 +134,7 @@ class Record_Quality_STR{
 public:
     UINT32  id;                   
 	UINT32	timestamp;
-	UINT16  meterNo;
+	UINT32  meterNo;
 	UINT16  flowPointIdx;
 	float   flowPoint;         //流量(m3/h)
 	UINT16  totalFlag;         //总量检定标志(1:总量检定  0:分量检定)
